[Decode] Fix MediaSolo PF for 8K and resolution change tests
1. Refactored MediaSolo Gfx address management with garbage collector and address check.
2. Fix incorrect page table entry idx issue in Solo Aub writing.
3. Fix history Solo Aub page table entry not cleaned properly during media reset.
4. Use 64-bit MediaSolo Gfx address.

+++ media_common/linux/common/os/mos_os_specific.h
@@ -361,7 +361,7 @@ struct _MOS_SPECIFIC_RESOURCE
     MOS_FORMAT          Format;
     int32_t             iCount;
     int32_t             iAllocationIndex[MOS_GPU_CONTEXT_MAX];
-    uint32_t            dwGfxAddress;
+    uint64_t            dwGfxAddress;
     uint8_t             *pData;
     const char          *bufname;
     uint32_t            isTiled;
